requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=10.0.0

# Data & Config
PyYAML>=6.0.1
pandas>=2.2.0
albumentations>=1.4.0

# Experiment Tracking
mlflow>=2.11.0
tensorboard>=2.16.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9

# Utilities
tqdm>=4.66.0
loguru>=0.7.2
rich>=13.7.0

# Dev / Testing
pytest>=8.1.0
pytest-cov>=5.0.0
ruff>=0.4.0
